+++ resources/views/admin/housetypes/edit.blade.php
@@ -35,7 +35,7 @@
             {!! Form::model($houseTypes, ['method'=>'PATCH', 'action'=> ['AdminHouseTypesController@update', $houseTypes->id], 'files' => true]) !!}
             <div class="form-group">
                 {!! Form::label('development_id', 'Development Name:')!!}
-                {!! Form::select('development_id', [''=>'Choose Development'] + $developments, null, ['class'=>'form-control selectHouseType']) !!}
+                {!! Form::select('development_id', [''=>'Choose Development'] + $developments, null, ['class'=>'form-control select']) !!}
             </div>
 
             <div class="form-group">
@@ -45,11 +45,11 @@
 
             <div class="form-group">
                 {!! Form::label('house_type_desc', 'House Type Description:')!!}
-                {!! Form::text('house_type_desc', null, ['class'=>'form-control']) !!}
+                {!! Form::textarea('house_type_desc', null, ['class'=>'form-control', 'rows' => 5, 'cols' => 40]) !!}
             </div>
 
             <div class="form-group">
-                {!! Form::label('floor_plan', 'Floor Plan Image:')!!}
+                {!! Form::label('flor_plan', 'Floor Plan Image:')!!}
                 {!! Form::file('floor_plan', null, ['class'=>'form-control']) !!}
             </div>
 
@@ -64,24 +64,58 @@
             {!! Form::close() !!}
 
 
-            {!! Form::open(['method'=>'DELETE', 'action'=> ['AdminHouseTypesController@destroy', $houseTypes->id]]) !!}
+            {!! Form::open(['method'=>'DELETE', 'action'=> ['AdminHouseTypesController@destroy', $houseTypes->id], 'id' => 'confirm_delete_'.$houseTypes->id]) !!}
             <div class="form-group">
-                {!! Form::submit('Delete House Type', ['class'=>'btn btn-danger col-sm-6']) !!}
+                {!! Form::submit('Delete House Type', ['class'=>'btn btn-danger col-sm-6', 'onclick'=>'confirmDelete(' .$houseTypes->id .')']) !!}
             </div>
             {!! Form::close() !!}
 
+
         </div>
     </div>
-    <div class="row">
-        @include('includes.form_error')
-    </div>
+    {{--<div class="row">--}}
+        {{--@include('includes.form_error')--}}
+    {{--</div>--}}
 
 @endsection
 
 @section('script')
     <script>
         $(document).ready(function() {
-            $('.selectHouseType').select2();
+            $('.select').select2();
         });
+
+        function confirmDelete(id) {
+            event.preventDefault();
+
+            swal({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                type: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!',
+                cancelButtonText: 'No, cancel!',
+                buttonsStyling: true
+            }).then((result) => {
+                if (result.value) {
+                swal({
+                    title: 'Deleted!',
+                    text: 'Your file has been deleted.',
+                    type: 'success',
+                    showConfirmButton: false,
+                    timer: 3000
+                }).then(function () {
+                    $("#confirm_delete_"+id).off("submit").submit()
+                })
+                // $("#confirm_delete_"+id).off("submit").submit()
+                // result.dismiss can be 'cancel', 'overlay',
+                // 'close', and 'timer'
+            } else if (result.dismiss === 'cancel') {
+
+            }
+        })
+        }
     </script>
 @endsection
