+++ resources/views/admin/plots/edit.blade.php
@@ -38,7 +38,7 @@
             {!! Form::model($plot, ['method'=>'PATCH', 'action'=> ['AdminPlotsController@update', $plot->id]]) !!}
             <div class="form-group">
                 {!! Form::label('development_id', 'Development:')!!}
-                {!! Form::select('development_id', [''=>'Choose Development'] + $developments, null, ['class'=>'form-control selectPlot']) !!}
+                {!! Form::select('development_id', [''=>'Choose Development'] + $developments, null, ['class'=>'form-control selectPlot select']) !!}
             </div>
             <div class="form-group">
                 {!! Form::label('plot_name', 'Plot Name:')!!}
@@ -46,7 +46,7 @@
             </div>
             <div class="form-group">
                 {!! Form::label('house_type', 'House Type:')!!}
-                {!! Form::select('house_type', [''=>'Choose House Type'] + $houseType, null, ['class'=>'form-control selectPlot']) !!}
+                {!! Form::select('house_type', [''=>'Choose House Type'] + $houseType, null, ['class'=>'form-control selectPlot select']) !!}
             </div>
             <div class="form-group">
                 {!! Form::label('sqft', 'SqFt:')!!}
@@ -58,7 +58,8 @@
             </div>
             <div class="form-group">
                 {!! Form::label('status', 'Status:')!!}
-                {!! Form::text('status', null, ['class'=>'form-control']) !!}
+                {!! Form::text('status_disabled', $plot->status, ['class'=>'form-control', 'disabled']) !!}
+                {!! Form::text('status', null, ['class'=>'form-control hidden']) !!}
             </div>
 
             <div class="form-group">
@@ -67,24 +68,56 @@
             {!! Form::close() !!}
 
 
-            {!! Form::open(['method'=>'DELETE', 'action'=> ['AdminPlotsController@destroy', $plot->id]]) !!}
+            {!! Form::open(['method'=>'DELETE', 'action'=> ['AdminPlotsController@destroy', $plot->id], 'id' => 'confirm_delete_'.$plot->id]) !!}
             <div class="form-group">
-                {!! Form::submit('Delete Plot', ['class'=>'btn btn-danger col-sm-6']) !!}
+                {!! Form::submit('Delete Plot', ['class'=>'btn btn-danger col-sm-6','onclick'=>'confirmDelete(' .$plot->id .')']) !!}
             </div>
             {!! Form::close() !!}
-
         </div>
     </div>
 
-    <div class="row">
-        @include('includes.form_error')
-    </div>
+    {{--<div class="row">--}}
+        {{--@include('includes.form_error')--}}
+    {{--</div>--}}
 
 @endsection
 @section('script')
     <script>
         $(document).ready(function(){
-            $('.selectPlot').select2();
+            $('.select').select2();
         });
+
+        function confirmDelete(id) {
+            event.preventDefault();
+
+            swal({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                type: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!',
+                cancelButtonText: 'No, cancel!',
+                buttonsStyling: true
+            }).then((result) => {
+                if (result.value) {
+                swal({
+                    title: 'Deleted!',
+                    text: 'Plot has been deleted.',
+                    type: 'success',
+                    showConfirmButton: false,
+                    timer: 3000
+                }).then(function () {
+                    $("#confirm_delete_"+id).off("submit").submit()
+                })
+                // $("#confirm_delete_"+id).off("submit").submit()
+                // result.dismiss can be 'cancel', 'overlay',
+                // 'close', and 'timer'
+            } else if (result.dismiss === 'cancel') {
+
+            }
+        })
+        }
     </script>
 @endsection
